Move to SceneDelegate
